refactor(desktop): organize chat attachment hooks
