+++ desktop/src/components/workspace/chat/ChatView.tsx
@@ -17,7 +17,7 @@ import {
 } from "@/hooks/chat/use-active-chat-session-selectors";
 import { useChatAvailabilityState } from "@/hooks/chat/use-chat-availability-state";
 import { useChatDockInset } from "@/hooks/chat/ui/use-chat-dock-inset";
-import { useChatPromptAttachments } from "@/hooks/chat/use-chat-prompt-attachments";
+import { useChatPromptAttachments } from "@/hooks/chat/ui/use-chat-prompt-attachments";
 import { useCloudWorkspacePolling } from "@/hooks/chat/use-cloud-workspace-polling";
 import { useComposerDockSlots } from "@/hooks/chat/ui/use-composer-dock-slots";
 import { useQueuedPromptEditStatus } from "@/hooks/chat/use-queued-prompt-edit";
