+++ desktop/src/hooks/chat/ui/use-chat-file-mention-search.ts
