+++ desktop/src/hooks/chat/ui/use-chat-prompt-attachments.ts
@@ -1,7 +1,7 @@
 import { useCallback } from "react";
 import type { PromptCapabilities } from "@anyharness/sdk";
 import { canAttachPromptContent } from "@/lib/domain/chat/composer/prompt-content";
-import { usePromptAttachments } from "@/hooks/chat/use-prompt-attachments";
+import { usePromptAttachments } from "@/hooks/chat/ui/use-prompt-attachments";
 import { useUserPreferencesStore } from "@/stores/preferences/user-preferences-store";
 
 export type PromptAttachmentController = ReturnType<typeof usePromptAttachments> & {
