+++ desktop/src/hooks/chat/ui/use-prompt-attachments.ts
